+++ storage/storage.go
@@ -3,8 +3,6 @@ package storage
 
 import (
 	"context"
-	"errors"
-	"time"
 
 	"github.com/jessepeterson/kmfddm/ddm"
 )
@@ -171,33 +169,6 @@ type StatusValuesRetriever interface {
 	// RetrieveStatusErrors retrieves the collected errors for enrollmentIDs.
 	RetrieveStatusValues(ctx context.Context, enrollmentIDs []string, pathPrefix string) (map[string][]StatusValue, error)
 }
-
-// StoredStatusReport represents a stored status report by StoreDeclarationStatus.
-type StoredStatusReport struct {
-	Raw       []byte    // the raw JSON bytes of the status report
-	Timestamp time.Time // the date the status report was saved
-	StatusID  string    // optional unique identifier of report. defined when report was saved.
-	Index     int       // optional "index" for this enrollment's status reports.
-}
-
-// StatusReportQuery specifies search criteria for finding specific status reports for enrollments.
-type StatusReportQuery struct {
-	EnrollmentID string
-	StatusID     *string
-	Index        *int
-}
-
-// Valid performs basic sanity checks for querying for status reports.
-func (q StatusReportQuery) Valid() error {
-	if q.EnrollmentID == "" {
-		return errors.New("missing enrollment ID")
-	} else if (q.StatusID == nil || *q.StatusID == "") && q.Index == nil {
-		return errors.New("status ID and index cannot both be empty")
-	}
-	return nil
-}
-
-var ErrStatusReportNotFound = errors.New("status report not found")
 
 type StatusReportRetriever interface {
 	RetrieveStatusReport(ctx context.Context, q StatusReportQuery) (*StoredStatusReport, error)
